Feat: Started moving clock to module
