+++ packages/protocol/engine/clock/clock.go
@@ -2,68 +2,27 @@ package clock
 
 import (
 	"time"
+
+	"github.com/iotaledger/goshimmer/packages/core/module"
 )
 
 // Clock is a clock that is used to derive some Time parameters from the Tangle.
-type Clock struct {
-	Events *Events
+type Clock interface {
+	// Events returns a dictionary that contains all events that are triggered by the Clock.
+	Events() *Events
 
-	lastAccepted  *timeUpdate
-	lastConfirmed *timeUpdate
-}
+	// AcceptedTime returns the time of the last accepted Block.
+	AcceptedTime() time.Time
 
-// New creates a new Clock with the given genesisTime.
-func New() (clock *Clock) {
-	return &Clock{
-		lastAccepted:  &timeUpdate{},
-		lastConfirmed: &timeUpdate{},
+	// RelativeAcceptedTime returns the real-time adjusted version of the time of the last accepted Block.
+	RelativeAcceptedTime() time.Time
 
-		Events: NewEvents(),
-	}
-}
+	// ConfirmedTime returns the time of the last confirmed Block.
+	ConfirmedTime() time.Time
 
-// AcceptedTime returns the Time of the last accepted Block.
-func (c *Clock) AcceptedTime() (acceptedTime time.Time) {
-	return c.lastAccepted.Time()
-}
+	// RelativeConfirmedTime returns the real-time adjusted version of the time of the last confirmed Block.
+	RelativeConfirmedTime() time.Time
 
-// SetAcceptedTime sets the Time of the last accepted Block.
-func (c *Clock) SetAcceptedTime(acceptedTime time.Time) (updated bool) {
-	now := time.Now()
-	if updated = c.lastAccepted.Update(now, acceptedTime); updated {
-		c.Events.AcceptanceTimeUpdated.Trigger(&TimeUpdateEvent{
-			NewTime:    acceptedTime,
-			UpdateTime: now,
-		})
-	}
-
-	return
-}
-
-// RelativeAcceptedTime returns the real-Time adjusted version of the Time of the last accepted Block.
-func (c *Clock) RelativeAcceptedTime() (relativeAcceptedTime time.Time) {
-	return c.lastAccepted.RelativeTime()
-}
-
-// ConfirmedTime returns the Time of the last confirmed Block.
-func (c *Clock) ConfirmedTime() (confirmedTime time.Time) {
-	return c.lastConfirmed.Time()
-}
-
-// SetConfirmedTime sets the Time of the last confirmed Block.
-func (c *Clock) SetConfirmedTime(confirmedTime time.Time) (updated bool) {
-	now := time.Now()
-	if updated = c.lastConfirmed.Update(now, confirmedTime); updated {
-		c.Events.ConfirmedTimeUpdated.Trigger(&TimeUpdateEvent{
-			NewTime:    confirmedTime,
-			UpdateTime: now,
-		})
-	}
-
-	return
-}
-
-// RelativeConfirmedTime returns the real-Time adjusted version of the Time of the last confirmed Block.
-func (c *Clock) RelativeConfirmedTime() (relativeConfirmedTime time.Time) {
-	return c.lastConfirmed.RelativeTime()
+	// Interface embeds the required methods of the module.Interface.
+	module.Interface
 }
