+++ packages/protocol/engine/clock/events.go
@@ -8,11 +8,11 @@ import (
 
 // Events is a collection of events that can be triggered by the Clock.
 type Events struct {
-	// AcceptanceTimeUpdated is triggered when a Acceptance Tangle Time advances.
-	AcceptanceTimeUpdated *event.Event1[*TimeUpdateEvent]
+	// AcceptanceTimeUpdated is an Event with the following signature func(newTime time.Time, now time.Time).
+	AcceptanceTimeUpdated *event.Event2[time.Time, time.Time]
 
 	// AcceptanceTimeUpdated is triggered when a Acceptance Tangle Time advances.
-	ConfirmedTimeUpdated *event.Event1[*TimeUpdateEvent]
+	ConfirmedTimeUpdated *event.Event2[time.Time, time.Time]
 
 	// LinkableCollection is a generic trait that allows to link multiple collections of events together.
 	event.Group[Events, *Events]
@@ -21,8 +21,8 @@ type Events struct {
 // NewEvents contains the constructor of the Events object (it is generated by a generic factory).
 var NewEvents = event.CreateGroupConstructor(func() (newEvents *Events) {
 	return &Events{
-		ConfirmedTimeUpdated:  event.New1[*TimeUpdateEvent](),
-		AcceptanceTimeUpdated: event.New1[*TimeUpdateEvent](),
+		ConfirmedTimeUpdated:  event.New2[time.Time, time.Time](),
+		AcceptanceTimeUpdated: event.New2[time.Time, time.Time](),
 	}
 })
