+++ packages/protocol/engine/clock/tangletime/timeupdate.go
@@ -1,4 +1,4 @@
-package clock
+package tangletime
 
 import (
 	"fmt"
